Share NATS globals across package
