+++ tests/unit/test_send_to_prism.py
@@ -4,7 +4,7 @@ import logging
 import aiohttp
 import pytest
 
-import examples.social_graph_bot as sg
+import examples.social_graph_bot as bot_mod
 
 
 class DummySession:
@@ -31,25 +31,25 @@ class TimeoutSession:
 
 @pytest.mark.asyncio
 async def test_send_to_prism_client_error(monkeypatch, caplog):
-    monkeypatch.setattr(sg.aiohttp, "ClientSession", lambda: DummySession())
+    monkeypatch.setattr(bot_mod.aiohttp, "ClientSession", lambda: DummySession())
     with caplog.at_level(logging.WARNING):
-        await sg.send_to_prism({"x": 1})
+        await bot_mod.send_to_prism({"x": 1})
     assert any("ClientError" in r.getMessage() for r in caplog.records)
 
 
 @pytest.mark.asyncio
 async def test_send_to_prism_timeout(monkeypatch, caplog):
-    monkeypatch.setattr(sg.aiohttp, "ClientSession", lambda: TimeoutSession())
+    monkeypatch.setattr(bot_mod.aiohttp, "ClientSession", lambda: TimeoutSession())
     with caplog.at_level(logging.WARNING):
-        await sg.send_to_prism({"x": 1})
+        await bot_mod.send_to_prism({"x": 1})
     assert any("TimeoutError" in r.getMessage() for r in caplog.records)
 
 
 @pytest.mark.asyncio
 async def test_send_to_prism_timeout_message(monkeypatch, caplog):
     """Ensure the timeout warning includes the expected text."""
-    monkeypatch.setattr(sg.aiohttp, "ClientSession", lambda: TimeoutSession())
+    monkeypatch.setattr(bot_mod.aiohttp, "ClientSession", lambda: TimeoutSession())
     with caplog.at_level(logging.WARNING):
-        await sg.send_to_prism({"x": 1})
+        await bot_mod.send_to_prism({"x": 1})
     messages = [rec.getMessage() for rec in caplog.records]
     assert any("TimeoutError sending data to Prism" in msg for msg in messages)
